Float notes; viewmodels updates; form files unification

+++ WD.Data/Models/FinalNote.cs
@@ -3,7 +3,7 @@
     public partial class FinalNote
     {
         public int NoteId { get; set; }
-        public short Note { get; set; }
+        public float Note { get; set; }
         public int StudentId { get; set; }
         public int ClassId { get; set; }
